Refactor api types
