+++ api.ts
@@ -1,36 +1,6 @@
-import { SampleUser, UserData } from "@/types/user";
-import { FolderData, SampleFolder } from "@/types/folder";
-import { Response, SampleFolderResponse } from "@/types/response";
-import { LinkData } from "@/types/link";
 import { CODEIT_BASE_URL } from "@/constants";
 import { FormValues } from "./common/Auth/Form";
-
-export interface Params {
-  [key: string]: number | undefined;
-}
-
-export async function getUser(): Promise<SampleUser> {
-  const response = await fetch(`${CODEIT_BASE_URL}/sample/user`);
-  if (!response.ok) {
-    throw new Error("잘못된 요청입니다.");
-  }
-
-  const user: SampleUser = await response.json();
-
-  return user;
-}
-
-export async function getFolder(): Promise<SampleFolder> {
-  const response = await fetch(`${CODEIT_BASE_URL}/sample/folder`);
-  if (!response.ok) {
-    throw new Error("잘못된 요청입니다.");
-  }
-
-  const data: SampleFolderResponse = await response.json();
-  const { folder } = data;
-
-  return folder;
-}
+import { FolderData, LinkData, Params, Response, UserData } from "./types/api";
 
 export async function getUserById({ userId }: Params): Promise<UserData> {
   const response = await fetch(`${CODEIT_BASE_URL}/users/${userId}`);
