--- types/response/index.ts
@@ -1,9 +1,0 @@
-import { SampleFolder } from "@/types/folder";
-
-export interface SampleFolderResponse {
-  folder: SampleFolder;
-}
-
-export interface Response<Data> {
-  data: Data[];
-}
